completion of Modal component
